- added FindTexture() to iLoaderContext - cosmetic changes in SyntaxService node path output - cosmetic changes in message box display - a bad XML token emits now a warning, not an error - ParseMixmode() is smarter; it detects when more than 1 exclusive   mode is specified - removed gl3d-ext from msvcgen required.mak - commented out "#error disable exceptions" in opcode
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@11022 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/cssys/win32/win32.cpp
@@ -478,7 +478,7 @@ bool csPlatformStartup(iObjectRegistry* r)
   char installDir[MAX_PATH];
   csGetInstallPath(installDir, sizeof(installDir));
   ToLower (installDir, installDir);
-  int idlen = strlen(installDir);
+  size_t idlen = strlen(installDir);
   idlen--;
   installDir[idlen] = 0;
